RDKTV-6079: USB upgrade failure error is shown
Reason for change: Remove the copy of the USB firmware file to the local download
directory on the device to be programmed. It adds about 10 seconds to the
programming time. The device will now be programmed directly from USB.

Test Procedure: Ability to upgrade device firmware from USB.

Risks: Low

Signed-off-by: Kirk Davis <kirk_davis@cable.comcast.com>
Change-Id: I74a7a98d536d0fba1cafafaab0afc60f13b6c0e8