LLAMA-3850 : Only bump maintenance time from xconf
Reason for change: The getMaintenanceStartTime script gets called from
numerous places. It is possible that a device that wakes up from sleep
can wake up within 4hrs of the maintenance window. When that happens,
the window time will be bumped by 24hrs. This change will remove the 24hr
bump from the getMaintenanceStartTime and keep the DCMscript_maintenance
script change, so the value will ony be changed once per maintenance cycle.

Test Procedure: Verify that if you wake up the device within 4hrs of the
maintenance window, that the value returned from getMaintenanceStartTime
returns the original value returned from DCM (unless it's in the past)

Risks: Low

Signed-off-by: Scott Angel <scott_angel@comcast.com>

LLAMA-3850 : Store the calculated epoch time for later use

Reason for change: We were only sotring the hr:min:sec in the start
time file, but this ends up losing the context when we request the
maintenance start time later and we need to recalculte. Now we are
able to pull the calculated time when AS requests via the plugin.

Test Procedure: Verify that if you wake up the device within 4hrs of the
maintenance window, that the value returned from getMaintenanceStartTime
returns the original value returned from DCM (unless it's in the past)

Risks: Low

Change-Id: Ifdc0b88d4c223de918d372fa59369f28d187b18a
Signed-off-by: Scott Angel <scott_angel@comcast.com>