RDKTV-14900 : Fix error on empty log folder
Reason for change: In RDKTV-14002, we stopped trying to tar up an empty
directory. In fixing the above, the code now works as designed and
generates an error when the this situation happens. This is not ideal
since we try to upload the logs quite frequently. So rather than post
an error if the log directory is empty, we'll just say the log upload
task is complete and move on.

Test Procedure: Verify that the 2nd log upload (from SOLICITED
maintenance) does not result in a MAINT_LOGUPLOAD_ERROR.

Risks: None

Signed-off-by: Scott Angel <scott_angel@comcast.com>
(cherry picked from commit d69bfe18a630ca39da3b89834232427253ac04be)
Change-Id: I928a68adf359f6bceeb2434e66c3827fedb278ae