DELIA-55499 : Add ability to mount SD card with journal
Reason for change: Sky is seeing XiOne devices that are booting
up with a corrupted SD card. This change adds the ability for a device
to create the SD card partitions with journaling support so it
has a higher probablity of recovering from disk errors.

Test Procedure:
On devices where SD_CARD_USE_JOURNAL is not defined (or not set to "true")
in /etc/device.properties. There should be no change. The SD card should
create the partitions and mount mmcblk0p1 and mmcblk0p2 with journal disabled
as verified in the /opt/logs/mount_logs.txt log file.

Devices that have SD_CARD_USE_JOURNAL set to "true" will create and mount the
mmcblk0p1 and mmcblk0p2 partitions with journal enabled as verified in the
/opt/logs/mount_log.txt file.

*NOTE* If your device has already mounted the SD card in the past, you will
need to factory reset your device for it to re-partition the SD card to enable
journaling.

Risks: Med

Signed-off-by: Scott Angel <scott_angel@comcast.com>
(cherry picked from commit 698dcc53baf054eebef7a265ad879442947cf6f6)
Change-Id: I65d9d190b311d6d61b3fee4fa15bd0b1a4a9b5be